Fixing register value out of range error

+++ src/com/intelligt/modbus/jlibmodbus/Modbus.java
@@ -290,7 +290,7 @@ final public class Modbus {
      * @return "true" if register value is correct, else "false".
      */
     static public boolean checkRegisterValue(int value) {
-        return checkRange((short)value, 0, Modbus.MAX_REGISTER_VALUE);
+        return checkRange(value, 0, Modbus.MAX_REGISTER_VALUE);
     }
 
     /**
